+++ src/mbe_automation/dynamics/harmonic/thermodynamics.py
@@ -215,10 +215,10 @@ def _hybrid_derivative(x: npt.NDArray[np.float64], y: npt.NDArray[np.float64]) -
         d_dx = d_dx_finite_diff
         
     else:
-        d_dx = np.zeros_like(x)
+        d_dx = np.zeros_like(x, dtype=np.float64)
         d_dx[0] = d_dx_finite_diff[0]
         d_dx[-1] = d_dx_finite_diff[-1]
-        d_dx[1:-1] = d_dx_cspline[1:-1]        
+        d_dx[1:-1] = d_dx_cspline[1:-1]
 
     return d_dx
